// src/types/game.ts
export type PlayerHand = { [key: number /* int */]: Card }
export type HasId = any
export interface Meld {
  id: number /* int */
  rank: Rank
  cards: Card[]
  wildCount: number /* int */
}
export interface Canasta {
  id: number /* int */
  rank: Rank
  cards: Card[]
  count: number /* int */
  natural: boolean
}
export interface Team {
  score: number /* int */
  melds: Meld[]
  canastas: Canasta[]
  goneDown: boolean
  canGoOut: boolean
  RedThrees: Card[]
}
export interface GameConfig {
  RandomTeamOrder: boolean
}
export type GameOption = any

//////////
// source: cards.go

export type Suit = number /* int */
export const Hearts: Suit = 0
export const Diamonds: Suit = 1
export const Clubs: Suit = 2
export const Spades: Suit = 3
export type Rank = number /* int */
export const Four = 0
export const Five = 1
export const Six = 2
export const Seven = 3
export const Eight = 4
export const Nine = 5
export const Ten = 6
export const Jack = 7
export const Queen = 8
export const King = 9
export const Ace = 10
export const Two = 11
export const Joker = 12
export const Three = 13
export const Wild = 14
export interface Card {
  id: number /* int */
  suit: Suit
  rank: Rank
}
// ...
export interface ClientState {
  deckCount: number /* int */
  discardCount: number /* int */
  discardTopCard?: Card // Pointer so we can send nil when pile is empty
  name: string
  hand: PlayerHand
  hasFoot: boolean
  players: OtherPlayerState[]
  ourScore: number /* int */
  ourMelds: Meld[]
  ourCanastas: Canasta[]
  ourRedThrees: Card[]
  otherScore: number /* int */
  otherMelds: Meld[]
  otherCanastas: Canasta[]
  otherRedThrees: Card[]
  goOutRequestPending: boolean // Is there an active permission request?
  isGoOutRequester: boolean // Is this player the one who asked?
  isGoOutPartner: boolean // Is this player the one who needs to respond?
}
export interface OtherPlayerState {
  name: string
  handLength: number /* int */
  hasFoot: boolean
}